+++ tests/testthat/resources/glue.R
@@ -1,5 +1,11 @@
 
 glue(
-  "Say hello: { hello::world() }",
-  "Say goodbye: { goodbye::world() }"
+  "{ library(A) }",
+  "{ library(B) }"
+)
+
+glue(
+  "< library(C) >",
+  .open  = "<",
+  .close = ">"
 )